Problem component
Creates a Problem component and shares it between the scoreboard and problem page.
Better reuse/consistency and opportunity to use it elsewhere in the future.
